remove test_lac
