+++ test/dygraph_to_static/check_approval.py
@@ -130,7 +130,6 @@ ALLOW_LIST: dict[type[Diagnostic], list[str]] = {
         "test_error.py",
         "test_op_attr.py",
         "test_se_resnet.py",
-        "test_lac.py",
         "test_convert_call.py",
         "test_local_cast.py",
         "test_origin_info.py",
@@ -163,7 +162,6 @@ ALLOW_LIST: dict[type[Diagnostic], list[str]] = {
         "test_typing.py",
         "test_gradname_parse.py",
         "test_for_enumerate.py",
-        "test_lac.py",
         "test_save_load.py",
         "test_cinn.py",
         "test_declarative.py",
